requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data processing
pandas==2.1.4
numpy==1.24.4
pyarrow==14.0.1
polars==0.20.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
imbalanced-learn==0.11.0
shap==0.44.0
lime==0.2.0.1

# Deep Learning
tensorflow==2.15.0
torch==2.1.1

# Feature Store
feast==0.34.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Message Queue
kafka-python==2.0.2
paho-mqtt==1.6.1

# Monitoring & Observability
prometheus-client==0.19.0
grafana-api==1.0.3
mlflow==2.8.1

# Web UI
streamlit==1.28.2
plotly==5.17.0
dash==2.14.2

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
